Fix warnings for Flymake in GCC 4.5.
GCC 4.5 uses a different format for errors and warnings, they added a number
which points to where in the line the error occurred. Flymake treats this as
part of the error text, and flymake decides a line is a warning if the error
text starts with the word [Ww]arning. This commit adds a regex which fixes
this.